+++ test/fuzz/accounts/AccountV1/GetAccountValue.fuzz.t.sol
@@ -25,9 +25,9 @@ contract GetAccountValue_AccountV1_Fuzz_Test is AccountV1_Fuzz_Test {
     /*//////////////////////////////////////////////////////////////
                               TESTS
     //////////////////////////////////////////////////////////////*/
-    function testFuzz_Success_getAccountValue(uint128 spotValue) public {
+    function testFuzz_Success_getAccountValue(uint112 spotValue) public {
         // Given: "exposure" is strictly smaller than "maxExposure".
-        spotValue = uint128(bound(spotValue, 0, type(uint128).max - 1));
+        spotValue = uint112(bound(spotValue, 0, type(uint112).max - 1));
 
         // Set Spot Value of assets (value of stable1 is 1:1 the amount of stable1 tokens).
         depositTokenInAccount(accountExtension, mockERC20.stable1, spotValue);
