+++ test/fuzz/accounts/AccountV1/GetFreeMargin.fuzz.t.sol
@@ -45,7 +45,7 @@ contract GetFreeMargin_AccountV1_Fuzz_Test is AccountV1_Fuzz_Test {
         creditorStable1.setOpenPosition(address(accountExtension), openDebt);
 
         // Given: "exposure" is strictly smaller than "maxExposure".
-        collateralValue = uint128(bound(collateralValue, 0, type(uint128).max - 1));
+        collateralValue = uint112(bound(collateralValue, 0, type(uint112).max - 1));
 
         // Set Liquidation Value of assets (Liquidation value of token1 is 1:1 the amount of token1 tokens).
         depositTokenInAccount(accountExtension, mockERC20.stable1, collateralValue);
@@ -56,14 +56,14 @@ contract GetFreeMargin_AccountV1_Fuzz_Test is AccountV1_Fuzz_Test {
     function testFuzz_Success_getFreeMargin_CreditorIsSet_NonZeroFreeMargin(
         uint256 openDebt,
         uint256 fixedLiquidationCost,
-        uint128 collateralValue
+        uint112 collateralValue
     ) public {
         // No overflow of Used Margin.
         vm.assume(openDebt <= type(uint256).max - fixedLiquidationCost);
 
         // "exposure" is strictly smaller than "maxExposure" -> collateralValue < type(uint128).max.
         // Non zero free margin -> "collateralValue" bigger than "usedMargin".
-        collateralValue = uint128(bound(collateralValue, 1, type(uint128).max - 1));
+        collateralValue = uint112(bound(collateralValue, 1, type(uint112).max - 1));
         fixedLiquidationCost = bound(fixedLiquidationCost, 0, collateralValue - 1);
         fixedLiquidationCost = bound(fixedLiquidationCost, 0, type(uint96).max);
         openDebt = bound(openDebt, 0, collateralValue - fixedLiquidationCost - 1);
@@ -83,7 +83,7 @@ contract GetFreeMargin_AccountV1_Fuzz_Test is AccountV1_Fuzz_Test {
     function testFuzz_Success_getFreeMargin_CreditorIsSet_ZeroFreeMargin(
         uint256 openDebt,
         uint256 fixedLiquidationCost,
-        uint128 collateralValue
+        uint112 collateralValue
     ) public {
         // No overflow of Used Margin.
         fixedLiquidationCost = bound(fixedLiquidationCost, 0, type(uint256).max - openDebt);
@@ -91,9 +91,9 @@ contract GetFreeMargin_AccountV1_Fuzz_Test is AccountV1_Fuzz_Test {
         uint256 usedMargin = openDebt + fixedLiquidationCost;
 
         // Zero free margin -> "collateralValue" smaller or equal as "usedMargin".
-        collateralValue = uint128(bound(collateralValue, 0, usedMargin));
+        collateralValue = uint112(bound(collateralValue, 0, usedMargin));
         // "exposure" is strictly smaller than "maxExposure" -> collateralValue < type(uint128).max.
-        collateralValue = uint128(bound(collateralValue, 0, type(uint128).max - 1));
+        collateralValue = uint112(bound(collateralValue, 0, type(uint112).max - 1));
 
         // Set fixedLiquidationCost
         accountExtension.setFixedLiquidationCost(uint96(fixedLiquidationCost));
