+++ test/fuzz/accounts/AccountV1/GetLiquidationValue.fuzz.t.sol
@@ -28,9 +28,9 @@ contract GetLiquidationValue_AccountV1_Fuzz_Test is AccountV1_Fuzz_Test {
     /*//////////////////////////////////////////////////////////////
                               TESTS
     //////////////////////////////////////////////////////////////*/
-    function testFuzz_Success_getLiquidationValue(uint128 spotValue, uint8 liquidationFactor) public {
+    function testFuzz_Success_getLiquidationValue(uint112 spotValue, uint8 liquidationFactor) public {
         // Given: "exposure" is strictly smaller than "maxExposure".
-        spotValue = uint128(bound(spotValue, 0, type(uint128).max - 1));
+        spotValue = uint112(bound(spotValue, 0, type(uint112).max - 1));
 
         // Set Spot Value of assets (value of "stable1" is 1:1 the amount of "stable1" tokens).
         depositTokenInAccount(accountExtension, mockERC20.stable1, spotValue);
@@ -41,7 +41,7 @@ contract GetLiquidationValue_AccountV1_Fuzz_Test is AccountV1_Fuzz_Test {
         // Set Liquidation factor of "stable1" for "stable1" to "liquidationFactor".
         vm.prank(users.riskManager);
         registryExtension.setRiskParametersOfPrimaryAsset(
-            address(creditorStable1), address(mockERC20.stable1), 0, type(uint128).max, 0, liquidationFactor
+            address(creditorStable1), address(mockERC20.stable1), 0, type(uint112).max, 0, liquidationFactor
         );
 
         uint256 expectedValue = uint256(spotValue) * liquidationFactor / RiskModule.ONE_4;
